make: Add check target and include testify in vendor
First of all, this change introduces the "make check" target
which is running unit test.

The another change is using glide-vc with --use-glide-lock options
which is the only wat to prevent removing of testify library
(because it's not being detected by "glide list" which is useb by
glide-vc by default).