make: Add check target and include testify in vendor
First of all, this change introduces a "make check" target
which runs unit tests.

The other change is using glide-vc with --use-glide-lock options
which is the only way to prevent the removal of the testify library
(because it's not being detected by "glide list" which is used in
glide-vc by default).