Support parameterized cloudconfig and passing certs
This adds support for booting EC2 instances with cloudconfigs
parameterized with go templates. It also adds support for passing
certificates to the instances using KMS.

To test it you need to create a KMS Key:

```
$ aws kms --region=eu-central-1 create-key --description="aws test"
```

And get it into `example/cluster.yml`.

Then pass the certificates directory with the following certificates to
the operator:

```
certs/
├── apiserver-ca.pem
├── apiserver-key.pem
├── apiserver.pem
├── calico
│   ├── client-ca.pem
│   ├── client-key.pem
│   └── client.pem
├── etcd
│   ├── server-ca.pem
│   ├── server-key.pem
│   └── server.pem
├── worker-ca.pem
├── worker-key.pem
└── worker.pem
```

And finally pass the cloudconfig templates for master and worker nodes
to the operator.

For now, if you want to log in to the instance you need to add your
public SSH key in the cloudconfig templates.