Changes circle to fetch architect unauthed if token not present
We usually fetch architect's latest binary authenticated, to
avoid any issues with rate limiting.
However, we don't pass secrets to forked builds, for security
reasons, so forked builds can't fetch the architect binary at all.

This changeset makes the fetch fall back to unauthenticated if the
token is not present - this means that forked builds work, but may
be subjected to GitHub rate limiting.