create: pass cloudconfig through S3
To avoid the 16KB limit aws imposes on user-data.

This is how it works:

1. We upload the final cloudconfig to s3
2. We generate a "small cloudconfig" whose only task is fetching the
final cloudconfig from s3
3. We configure the instance to be able to access the s3 URI where the
final cloudconfig is stored
4. We start the instance with the "small cloudconfig"

At runtime, the instance will boot with the "small cloudconfig", then it
will fetch the final cloudconfig, and finally it will run
coreos-cloudinit again against the final cloudconfig.