service: rework roles/policies/instance profiles
We now name roles, policies and instance profiles in a per-cluster
basis.

Also, we treat already-existing roles or instance profiles as errors.
Reusing them doesn't really work. In any case, they should've been
removed when we deleted the cluster (for now we need to do it manually
since the delete operator is not implemented yet).