Adds region to bucket name
Towards https://github.com/giantswarm/aws-operator/issues/147

If we create a bucket in a particular region, and then try to create
clusters for the same customer in a different region, we re-use
the previous bucket, which fails as it is in the wrong region.

This changeset adds the region to the bucketname, so the buckets
are correct for each region.