Make worker docker volume size configurable via CR
Same docker volume size doesn't fit to all tenant cluster use cases. It
makes sense to be able to configure it per guest cluster. Earlier
implementation always defaulted to 100GB (which is the default now as
well). When the field is present, its value is used instead of default
one.