Wire v16patch1 & update k8scloudconfig
Kubernetes v1.11.1 (and also .2 and .3) apiserver has a memory leak.
There is a fix in apimachinery for it, but upstream release PR is
waiting for testing atm. To mitigate this issue for us in the meanwhile,
we built our own patched version of hyperkube image to enable k8s
v1.11.x clusters.