Improve legacy tenant subnet CIDR detection
When checking AWSConfig CR if there's a tenant subnet already allocated
in legacy CIDR field, improve the check so that if it's a proper CIDR
then use it and otherwise proceed with new subnet allocation.

This way we can either use empty string or "DEPRECATED" etc. in legacy
CIDR field for migration period.