terminate the master after detaching of volumes
once the volumes are detached from the master we can terminate the
instance.

the instance would be terminate by the cloudformation too, but only at
the very end. and since k8s keeps the lost master node around as long as
it is only "Stopped" we need to manually terminate the master.