Retain non-indexed resource name for first AZ resources (#1293)
* Retain non-indexed resource name for first AZ resources

In order to provide smooth migration path for old clusters and allow
theoretical possibility to increase number of AZs in the future,
existing resources names must not change, because CloudFormation cannot
recognize change of resource name. Instead it will create new resources
with new names next to old ones and that will end up in error due to
duplicate values, e.g. VPC subnet network range.

With the current implementation change in number of availability zones
is not possible, but in the future it could be if we enable use of
multiple network segments for one tenant cluster. Same rules for
CloudFormation apply still and hence the implementation is done so that
first instance of resource doesn't have indexes and following ones start
counting from 1.

* add multiple private subnets to etcd elb

* add private subnets to loadbalancer adapter

* fix private subnets for etcd elb

* fix master subnet and etcd elb security group

* fix autoscaler private subnets for multiaz

* Address review comments

* Update e2etemplates

* Update e2etemplates