Use key functions for CF template resource names
When generating indexed resource names to CloudFormation template, use
key functions for that so that adapter code is simpler when idx 0 and
idx 1 differ in logic and also so that it's easier to maintain resource
name pattern when there are references to it in different places. Route
table names and subnet names are commonly referenced from different
parts of template so using key function for this simplifies code quite a
bit while also making it a bit more robust.