Short circuit ELB metrics collection when no LBs present
During cluster creation there is a period of time when AWSConfig is
reconciled, cluster is in `Creating` state but no load balancers present
yet. Querying AWS API for loadbalancer tags fails with validation error
when there are no LoadBalancerNames in the request. Therefore short
circuit when no load balancers exist.