Add service to provide concurrent safe IPAM (#1548)
With Node Pools there is a need for performing network allocations from
multiple controllers concurrently. In order to make it correctly, start by
creating an implementation that performs inventory on existing networks and
persisting of allocated network within a lock so that there is no race
condition between callers.

Later with VOO the in-process mutex must be changed to be a distributed lock in
order to maintain correct behavior.