Fix timing dependent test logic in network allocation test
Dependency on timing in test execution is always a bad thing. In this case it
was originally there so that main test target doesn't need to be adapted for
test, but since there are CI environments where test execution is seriously
slow, no time.Sleep() might be appropriate for test in question.

Therefore this change adds optional preAllocateHook() in Allocate() function
start so that unit-test can set a callback that is then used for correct
signaling between concurrent goroutines.