Remove broken test that's too complicated for maintenance
As deemed in another PR: https://github.com/giantswarm/aws-operator/pull/1750
this test is very complicated and it's questionable if it's even correct once
it's not outright broken as such.

In this present form it's timing dependent on several places but even in the
referenced PR when the time.Sleep() was removed, it came up that even the test
premise validation part must be synchronized.

Therefore this is removed for now and if we ever face problems with the actual
`Allocate()` implementation, we can then reconsider resuming this test and
fixing it properly. For now it's not worth it.