Drop AZs that are not required by CRs (Cluster, MachineDeployment)
tccpazs resource maintains state of reserved availability zones for cluster and
persists this in context for operatorkit reconciliation round. This information
is then further used e.g. for CloudFormation stack creation / updates so that's
why this AZ set must be correct and why changes to it further ensure changes to
final cluster state.