Don't release once allocated AZs
Due to current behavior of AWS CloudFormation / ELB, we can't release AZs that
were once allocated for cluster. Reason for this is that when subnet for freed
AZ is detached from ELBs, the requester-managed network interface (ENI created
by ELB) is not deleted from corresponding subnet and therefore this subnet
cannot be deleted. This keeps allocated network range reserved and it cannot be
used for another AZ.

Thus for now the AZ allocation for cluster / node pool is permanent and cannot
be reverted.